+++ engine/musicbrainz.rb
@@ -79,23 +79,28 @@ module Onebox
             @data[:image_source] = image_url
             @data[:image_source_label] = URI.parse(image_url).host.downcase
           end
-        else
-          wikidata_image(WIKIDATA_TYPE_IMAGE)
-          if @data[:image].nil?
-            wikidata_image(WIKIDATA_TYPE_LOGO_IMAGE)
-          end
         end
       end
 
-      def wikidata_image(type)
+      def wikidata()
         wikidata = get_relations("url", ["wikidata"]).first
-        if wikidata
-          wikidata_url = wikidata["url"]["resource"]
-          @data[:image] = wikidata_image_url(wikidata_url, type)
-          if !@data[:image].nil?
-            @data[:image_source] = wikidata_url
-            @data[:image_source_label] = "Wikidata"
-          end
+        return nil if wikidata.nil?
+        url = wikidata["url"]["resource"]
+        data = wikidata_data(url)
+        wiki = data["sitelinks"]["enwiki"]
+        if wiki
+          @data[:wikipedia_url] = wiki["url"]
+          @data[:wikipedia_title] = wiki["title"]
+        end
+        wikidata_image(data, url, WIKIDATA_TYPE_IMAGE) if @data[:image].nil?
+        wikidata_image(data, url, WIKIDATA_TYPE_LOGO_IMAGE) if @data[:image].nil?
+      end
+
+      def wikidata_image(url, data, type)
+        @data[:image] = wikidata_image_url(data, type)
+        if !@data[:image].nil?
+          @data[:image_source] = url
+          @data[:image_source_label] = "Wikidata"
         end
       end