+++ engine/musicbrainz_series_onebox.rb
@@ -32,6 +32,7 @@ module Onebox
 
         disambiguation
         image
+        wikidata
 
         return @data
       end